complete sns-service 코드 최적화

+++ fastcampus/sns-service/src/main/java/com/fastcampus/sns/controller/PostController.java
@@ -68,7 +68,7 @@ public class PostController {
 
     @GetMapping("/{postId}/likes")
     @ResponseBody
-    public Response<Integer> likeCount(@PathVariable Integer postId, Authentication authentication) {
+    public Response<Long> likeCount(@PathVariable Integer postId, Authentication authentication) {
 
         return Response.success(postService.likeCount(postId));
     }
